+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 dev
 ---
 
+* Added new options: `max_length` and `ellipsis`. They allow to limit the
+  maximal length of notifications.
 * Added a new option: `escape_html`. It escapes the `<`, `>`, and `&` HTML
   characters in notification messages.